+++ elide-integration-tests/src/test/java/com/yahoo/elide/tests/AnyPolymorphismIT.java
@@ -1,0 +1,253 @@
+/*
+ * Copyright 2017, Yahoo Inc.
+ * Licensed under the Apache License, Version 2.0
+ * See LICENSE file in project root for terms.
+ */
+package com.yahoo.elide.tests;
+
+import com.jayway.restassured.RestAssured;
+import com.jayway.restassured.response.Response;
+import com.yahoo.elide.core.HttpStatus;
+import com.yahoo.elide.initialization.AbstractApiResourceInitializer;
+import com.yahoo.elide.utils.JsonParser;
+import org.testng.Assert;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Test;
+
+import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.equalTo;
+
+public class AnyPolymorphismIT extends AbstractApiResourceInitializer {
+    private static final String JSONAPI_CONTENT_TYPE = "application/vnd.api+json";
+    private final JsonParser jsonParser = new JsonParser();
+
+    @BeforeClass
+    public void setUp() {
+        //Create a tractor
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .body("{\"data\": {\"type\": \"tractor\", \"attributes\": {\"horsepower\": 102 }}}")
+                .post("/tractor")
+                .then()
+                .statusCode(HttpStatus.SC_CREATED);
+
+        //Create a smartphone
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .body("{\"data\": {\"type\": \"smartphone\", \"attributes\": {\"type\": \"android\" }}}")
+                .post("/smartphone")
+                .then()
+                .statusCode(HttpStatus.SC_CREATED);
+    }
+
+    @Test
+    public void testAny() {
+
+        String id1 = RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .body(jsonParser.getJson("/AnyPolymorphismIT/AddTractorProperty.json"))
+                .post("/property")
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_CREATED)
+                .extract()
+                .path("data.id");
+
+        String id2 = RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .body(jsonParser.getJson("/AnyPolymorphismIT/AddSmartphoneProperty.json"))
+                .post("/property")
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_CREATED)
+                .extract()
+                .path("data.id");
+
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .get("/property/" + id1)
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_OK)
+                .body("data.relationships.myStuff.data.type", equalTo("tractor"),
+                        "data.relationships.myStuff.data.id", equalTo("1"));
+
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .get("/property/" + id2)
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_OK)
+                .body("data.relationships.myStuff.data.type", equalTo("smartphone"),
+                        "data.relationships.myStuff.data.id", equalTo("1"));
+
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .get("/property/" + id1 + "?include=myStuff")
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_OK)
+                .body("included[0].type", equalTo("tractor"),
+                        "included[0].id", equalTo("1"),
+                        "included[0].attributes.horsepower", equalTo(102),
+                        "included.size()", equalTo(1));
+
+
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .get("/property/" + id2 + "?include=myStuff")
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_OK)
+                .body("included[0].type", equalTo("smartphone"),
+                        "included[0].id", equalTo("1"),
+                        "included[0].attributes.type", equalTo("android"),
+                        "included.size()", equalTo(1),
+                        "included[0].attributes.operatingSystem", equalTo("some dessert"));
+
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .get("/property")
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_OK)
+                .body("data.size()", equalTo(2));
+    }
+
+    @Test
+    public void testAnyupdate() {
+        String id = RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .body(jsonParser.getJson("/AnyPolymorphismIT/AddTractorProperty.json"))
+                .post("/property")
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_CREATED)
+                .body("data.relationships.myStuff.data.type", equalTo("tractor"))
+                .extract()
+                .path("data.id");
+
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .body("{\"data\": {\"id\": \"" + id + "\", \"type\": \"property\", \"relationships\": {\"myStuff\": {\"data\": {\"type\": \"smartphone\", \"id\": \"1\"}}}}}")
+                .patch("/property/" + id)
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_NO_CONTENT);
+
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .get("/property" + "/" + id)
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_OK)
+                .body("data.relationships.myStuff.data.type", equalTo("smartphone"));
+
+        //delete relation
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .body("{\"data\": {\"id\": \"" + id + "\", \"type\": \"property\", \"relationships\": {\"myStuff\": {\"data\": null}}}}")
+                .patch("/property/" + id)
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_NO_CONTENT);
+
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .get("/property/" + id)
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_OK)
+                .body("data.relationships.myStuff.data", equalTo(null));
+    }
+
+    @Test
+    public void testAnySubpaths() {
+        String id = RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .body(jsonParser.getJson("/AnyPolymorphismIT/AddTractorProperty.json"))
+                .post("/property")
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_CREATED)
+                .extract()
+                .path("data.id");
+
+        Response response = RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .get("/property?page[totals]")
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_OK)
+                .extract()
+                .response();
+
+        Integer collectionSize = response.path("data.size()");
+        Integer totalSize = response.path("meta.page.totalRecords");
+        Assert.assertEquals(collectionSize, totalSize);
+
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .get("/property/" + id + "/myStuff")
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_OK)
+                .body("data.attributes.horsepower", equalTo(102),
+                        "data.id", equalTo("1"));
+
+        //show that you can't get a relation by the interface type
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .get("/device")
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_NOT_FOUND)
+                .body("errors[0]", containsString("Unknown collection"));
+
+        //Filtering is not supported for these types.
+        RestAssured
+                .given()
+                .contentType(JSONAPI_CONTENT_TYPE)
+                .accept(JSONAPI_CONTENT_TYPE)
+                .get("/property/" + id + "?filter[tractor]=horsepower==103")
+                .then()
+                .assertThat()
+                .statusCode(HttpStatus.SC_BAD_REQUEST);
+    }
+}